Fix test wasn't execute with rename
