+++ spec/features/order_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.feature "Order::OrderItems", type: :feature do
+RSpec.feature 'Order', type: :feature do
 
   feature '注文の確認' do
     scenario '注文者は自分の注文を確認できる' do
@@ -39,7 +39,7 @@ RSpec.feature "Order::OrderItems", type: :feature do
       # 予約ページで情報を入力
       click_link('予約する')
       fill_in 'Customer name', with: 'customer'
-      select 'sample弁当', from: 'lunchbox[id]'
+      select 'sample弁当', from: 'order_item[lunchbox_id]'
 
       # 弁当発注の確定
       order.closed_at = Time.zone.local(2017, 2, 1)
@@ -56,6 +56,7 @@ RSpec.feature "Order::OrderItems", type: :feature do
 
   end
 
+
   feature '注文のキャンセル' do
     scenario '注文者は自分の注文をキャンセルできる' do
       lunchbox = create(:lunchbox)
@@ -64,7 +65,7 @@ RSpec.feature "Order::OrderItems", type: :feature do
 
       visit order_order_items_path(order)
       expect(page).to have_text(order_item.customer_name)
-      expect(page).to have_text("cancel")
+      expect(page).to have_text('cancel')
 
       click_link('cancel')
       expect(page).not_to have_text(order_item.customer_name)
@@ -83,9 +84,56 @@ RSpec.feature "Order::OrderItems", type: :feature do
       expect(page).to have_text(order_item.customer_name)
 
     end
+
   end
 
+  feature '注文の修正' do
+    scenario '注文者は自分の注文を修正できる' do
+      lunchbox = create(:lunchbox)
+      create(:lunchbox, name: 'sample弁当-上')
+      order = create(:order)
+      order_item = create(:order_item, order: order, lunchbox: lunchbox)
+      new_name = 'another_name'
+      new_lunchbox_name = 'sample弁当-上'
 
+      visit order_order_items_path(order)
+      expect(page).to have_text(order_item.customer_name)
+
+      # edit name
+      click_link(order_item.customer_name)
+      fill_in 'Customer name', with: new_name
+
+      # change lunchbox
+      expect(page).to have_select('order_item[lunchbox_id]',selected: 'sample弁当')
+      select new_lunchbox_name, from: 'order_item[lunchbox_id]'
+
+      # update confirm
+      click_button 'Update Order item'
+      expect(page).not_to have_text(order_item.customer_name)
+      expect(page).to have_text(new_name)
+
+      # confirm new order in edit page
+      click_link(new_name)
+      expect(page).to have_select('order_item[lunchbox_id]',selected: new_lunchbox_name)
+
+    end
+
+    scenario 'Order が締め切られている場合、注文者は自分の注文を編集できない' do
+      lunchbox = create(:lunchbox)
+      order = create(:order, closed_at: Time.zone.local(2017, 2, 1))
+      order_item = create(:order_item, order: order, lunchbox: lunchbox)
+
+      visit order_order_items_path(order)
+      expect(page).to have_text(order_item.customer_name)
+
+      # edit name
+      click_link(order_item.customer_name)
+
+      # トップページに戻される
+      expect(page).to have_text('予約する')
+    end
+
+  end
 
 end
